app/requirements.txt
dj-database-url==0.5.0
Django==2.1.5
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-extensions==2.1.6
django-heroku==0.3.1
Faker==1.0.4
graphviz==0.10.1
gunicorn==19.9.0
Jinja2==2.10
MarkupSafe==1.1.0
Pillow==5.4.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pydot==1.4.1
pydotplus==2.0.2
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
whitenoise==4.1.2
